+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `LibraryError` wrapper type to avoid exposing `libloading` types
 - Add `try_vk` macro, like `try`, but for `VulkanResult`
 - Add Manual Page to Builder Documentation
+- Add `#[inline]` to `VulkanResult` functions
 
 ### Changed
 - Change cargo feature representation in the README